requirements.txt
click==8.2.1
ghp-import==2.1.0
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
six==1.17.0
watchdog==6.0.0
flask
flask-wtf
wtforms
streamlit
python-dotenv
google-generativeai  # Para usar a API do Gemini (Google AI)
sentence-transformers # Para gerar os embeddings
faiss-cpu             # Para o vector store (alternativa: chromadb)